Added authorization to Paper
Changed author list to hold User objects, updating CRUD methods to respect this
Added getters and setters for attributes to User
Added fetch methods to user for other users to get info on a user
Updated Paper to require login for all CRUD methods. User can be passed through constructor or setter. Logged in user must be an author or admin to create, read, or update. Must be admin to delete.